Move the image plug-in to a separate file
