+++ eleventy.config.images.js
@@ -1,0 +1,23 @@
+const Image = require("@11ty/eleventy-img");
+
+async function imageShortcode(src, alt, sizes) {
+    let metadata = await Image(`./src${src}`, {
+        widths: [300, 800, null],
+        formats: ["avif", "jpeg"],
+        urlPath: "/images/",
+        outputDir: "./_site/images/"
+    });
+
+    let imageAttributes = {
+        alt,
+        sizes,
+        loading: "lazy",
+        decoding: "async"
+    };
+
+    return Image.generateHTML(metadata, imageAttributes);
+}
+
+module.exports = function(eleventyConfig) {
+    eleventyConfig.addNunjucksAsyncShortcode("EleventyImage", imageShortcode);
+}
